pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
discord-py = "^2.4.0"
youtube-dl = "^2021.12.17"
yt-dlp = "^2024.8.6"
pafy = "^0.5.5"
python-vlc = "^3.0.20123"
ffmpeg = "1.4"
ffmpeg-asyncio = "0.1.3"
ffmpeg4discord = "0.1.4"
spotipy = "2.24.0"
spotipy2 = "0.6.1"
pytube = "^15.0.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"